print computing of offset

+++ tracer/champsim_tracer.cpp
@@ -320,8 +320,8 @@ void MemoryWrite(VOID* addr, UINT32 index)
 // memory address stored in register 
 void findOffset (VOID* effectiveAddr, UINT32 regAddr)
 {
-    cout << "the effective address: " << effectiveAddr << endl ;
-    cout << hex << "the address stored in register: 0x" << regAddr << dec << endl ;
+    cout << "the effective address: " << (unsigned long long int) effectiveAddr << endl ;
+    cout << hex << "the address stored in register: 0x" << (unsigned long long int) regAddr << dec << endl ;
     curr_instr.offset = (unsigned long long int) effectiveAddr - (unsigned long long int) regAddr;
 }
